mod: introduce load_balancing module
This PR introduces a lot of stuff related to load balancing. It does
not make much sense to have it all in cluster.rs.
