+++ scylla-rust-wrapper/src/load_balancing.rs
@@ -1,0 +1,76 @@
+use std::sync::Arc;
+
+use scylla::policies::load_balancing::{
+    DefaultPolicyBuilder, LatencyAwarenessBuilder, LoadBalancingPolicy,
+};
+
+#[derive(Clone, Debug)]
+pub(crate) struct LoadBalancingConfig {
+    pub(crate) token_awareness_enabled: bool,
+    pub(crate) token_aware_shuffling_replicas_enabled: bool,
+    pub(crate) load_balancing_kind: Option<LoadBalancingKind>,
+    pub(crate) latency_awareness_enabled: bool,
+    pub(crate) latency_awareness_builder: LatencyAwarenessBuilder,
+}
+
+impl LoadBalancingConfig {
+    // This is `async` to prevent running this function from beyond tokio context,
+    // as it results in panic due to DefaultPolicyBuilder::build() spawning a tokio task.
+    pub(crate) async fn build(self) -> Arc<dyn LoadBalancingPolicy> {
+        let load_balancing_kind = self
+            .load_balancing_kind
+            // Round robin is chosen by default for cluster wide LBP.
+            .unwrap_or(LoadBalancingKind::RoundRobin);
+
+        let mut builder = DefaultPolicyBuilder::new().token_aware(self.token_awareness_enabled);
+        if self.token_awareness_enabled {
+            // Cpp-driver enables shuffling replicas only if token aware routing is enabled.
+            builder =
+                builder.enable_shuffling_replicas(self.token_aware_shuffling_replicas_enabled);
+        }
+
+        match load_balancing_kind {
+            LoadBalancingKind::DcAware { local_dc } => {
+                builder = builder.prefer_datacenter(local_dc).permit_dc_failover(true)
+            }
+            LoadBalancingKind::RackAware {
+                local_dc,
+                local_rack,
+            } => {
+                builder = builder
+                    .prefer_datacenter_and_rack(local_dc, local_rack)
+                    .permit_dc_failover(true)
+            }
+            LoadBalancingKind::RoundRobin => {}
+        }
+
+        if self.latency_awareness_enabled {
+            builder = builder.latency_awareness(self.latency_awareness_builder);
+        }
+        builder.build()
+    }
+}
+
+impl Default for LoadBalancingConfig {
+    fn default() -> Self {
+        Self {
+            token_awareness_enabled: true,
+            token_aware_shuffling_replicas_enabled: true,
+            load_balancing_kind: None,
+            latency_awareness_enabled: false,
+            latency_awareness_builder: Default::default(),
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub(crate) enum LoadBalancingKind {
+    RoundRobin,
+    DcAware {
+        local_dc: String,
+    },
+    RackAware {
+        local_dc: String,
+        local_rack: String,
+    },
+}
